some updates, mostly for scale controller stuff
